# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Brave Accounts Service",
        "contact": {}
    },
    "paths": {
        "/v2/accounts/password/finalize": {
            "post": {
                "description": "Complete the password setup process and return auth token.\nEither `publicKey`, `maskingKey` and `envelope` must be provided together,\nor `serializedRecord` must be provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Finalize password setup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + verification token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    },
                    {
                        "description": "Registration record",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RegistrationRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.PasswordFinalizeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v2/accounts/password/init": {
            "post": {
                "description": "Start the password setup process using OPAQUE protocol.\nIf `serializeResponse` is set to true, the `serializedResponse` field will be populated\nin the response, with other fields omitted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Initialize password setup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + verification token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Brave services key (if one is configured)",
                        "name": "Brave-Key",
                        "in": "header"
                    },
                    {
                        "description": "Registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.RegistrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.PasswordFinalizeResponse": {
            "description": "Response for password setup or change",
            "type": "object",
            "properties": {
                "authToken": {
                    "description": "Authentication token",
                    "type": "string"
                }
            }
        },
        "controllers.RegistrationRecord": {
            "description": "OPAQUE registration record for a new account",
            "type": "object",
            "properties": {
                "envelope": {
                    "description": "Envelope of registation record",
                    "type": "string"
                },
                "maskingKey": {
                    "description": "Masking key of registation record",
                    "type": "string"
                },
                "publicKey": {
                    "description": "Public key of registation record",
                    "type": "string"
                },
                "serializedRecord": {
                    "description": "Serialized registration record",
                    "type": "string"
                },
                "sessionName": {
                    "description": "Optional name of the new session",
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "controllers.RegistrationRequest": {
            "description": "Request to register a new account",
            "type": "object",
            "required": [
                "blindedMessage"
            ],
            "properties": {
                "blindedMessage": {
                    "description": "Serialized OPAQUE registration request",
                    "type": "string"
                },
                "serializeResponse": {
                    "description": "Whether to serialize the response into binary/hex",
                    "type": "boolean"
                }
            }
        },
        "controllers.RegistrationResponse": {
            "description": "Response for registering a new account",
            "type": "object",
            "properties": {
                "evaluatedMessage": {
                    "description": "Evaluated message of the OPAQUE registration response",
                    "type": "string"
                },
                "pks": {
                    "description": "PKS of the OPAQUE registration response",
                    "type": "string"
                },
                "serializedResponse": {
                    "description": "Serialized OPAQUE registration response",
                    "type": "string"
                }
            }
        },
...
        "util.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        }
    },
...
}